+++ src/main/java/com/owlmaddie/network/ServerPackets.java
@@ -264,20 +264,27 @@ public class ServerPackets {
 
         // Grab random adjective
         String randomAdjective = Randomizer.getRandomMessage(Randomizer.RandomType.ADJECTIVE);
-        String randomFrequency = Randomizer.getRandomMessage(Randomizer.RandomType.FREQUENCY);
+        String randomClass = Randomizer.getRandomMessage(Randomizer.RandomType.CLASS);
+        String randomAlignment = Randomizer.getRandomMessage(Randomizer.RandomType.ALIGNMENT);
+        String randomSpeakingStyle = Randomizer.getRandomMessage(Randomizer.RandomType.SPEAKING_STYLE);
 
+        // Generate random name parameters
+        String randomLetter = Randomizer.RandomLetter();
+        int randomSyllables = Randomizer.RandomNumber(5) + 1;
+
+        // Build the message
         StringBuilder userMessageBuilder = new StringBuilder();
-        userMessageBuilder.append("Please generate a " + randomFrequency + " " + randomAdjective);
-        userMessageBuilder.append(" character ");
+        userMessageBuilder.append("Please generate a ").append(randomAdjective).append(" character. ");
+        userMessageBuilder.append("This character is a ").append(randomClass).append(" class, who is ").append(randomAlignment).append(". ");
         if (entity.getCustomName() != null && !entity.getCustomName().getString().equals("N/A")) {
-            userMessageBuilder.append("named '").append(entity.getCustomName().getString()).append("' ");
+            userMessageBuilder.append("Their name is '").append(entity.getCustomName().getString()).append("'. ");
         } else {
-            userMessageBuilder.append("whose name starts with the letter '").append(Randomizer.RandomLetter()).append("' ");
-            userMessageBuilder.append("and uses ").append(Randomizer.RandomNumber(4) + 1).append(" syllables ");
+            userMessageBuilder.append("Their name starts with the letter '").append(randomLetter)
+                    .append("' and is ").append(randomSyllables).append(" syllables long. ");
         }
-        userMessageBuilder.append("and speaks in '" + userLanguage + "'" );
-        LOGGER.info(userMessageBuilder.toString());
+        userMessageBuilder.append("They speak in '").append(userLanguage).append("' with a ").append(randomSpeakingStyle).append(" style.");
 
+        LOGGER.info(userMessageBuilder.toString());
         chatData.generateMessage(userLanguage, player, "system-character", userMessageBuilder.toString(), false);
     }